+++ tests/test_assign_taxonomy.py
@@ -1275,11 +1275,40 @@ class MothurTaxonAssignerTests(TestCase):
         ])
         self.assertTrue(x_conf > 0.94)
 
+    def test_assignment_with_collision(self):
+        """Taxa with spaces should be distinct from taxa with underscores.
+        """
+        fd, id_to_taxonomy_fp = mkstemp(
+            prefix='MothurTaxonAssigner_',
+            suffix='.txt')
+        close(fd)
+
+        with open(id_to_taxonomy_fp, "w") as f:
+            f.write(mothur_id_to_taxonomy_collision)
+        self.params['id_to_taxonomy_fp'] = id_to_taxonomy_fp
+
+        with open(self.seq_fp1, "w") as f:
+            f.write(mothur_input_collision)
+
+        assigner = MothurTaxonAssigner(self.params)
+        result = assigner(self.seq_fp1)
+
+        spaces_lineage, _ = result['spaces_seq']
+        self.assertEqual(spaces_lineage[:4], [
+            'Bacteria', 'Proteo bacteria', 'Gammaproteobacteria',
+            'Vibr iona les'])
+
+        underscores_lineage, _ = result['underscores_seq']
+        self.assertEqual(underscores_lineage[:4], [
+            'Bacteria', 'Proteo bacteria', 'Gammaproteobacteria',
+            'Vibr_iona_les'])
+
     def test_assignment_to_file(self):
         fd, output_fp = mkstemp(
             prefix='MothurTaxonAssigner_',
             suffix='.txt')
         close(fd)
+        self._paths_to_clean_up.append(output_fp)
         assigner = MothurTaxonAssigner(self.params)
 
         result = assigner(self.seq_fp1, result_path=output_fp)
@@ -1821,6 +1850,23 @@ xxxxxx	Bacteria; Proteo bacteria; Gammaproteobacteria; Pseudomonadales; Pseudomo
 AB004748	Bacteria; Proteo bacteria; Gammaproteobacteria; Enterobacter iales; Enterobacteriaceae; Enterobacter
 AB000278	Bacteria; Proteo bacteria; Gammaproteobacteria; Vibrionales; Vibrionaceae; Photobacterium
 AB000390	Bacteria; Proteo bacteria; Gammaproteobacteria; Vibrionales; Vibrionaceae; Vibrio
+"""
+
+mothur_id_to_taxonomy_collision = \
+    """X67228	Bacteria; Proteo bacteria; Alphaproteobacteria; Rhizobiales; Rhizobiaceae; Rhizobium
+X73443	Bacteria; Firmicutes; Clostridia; Clostridiales; Clostridiaceae; Clostridium
+AB004750	Bacteria; Proteo bacteria; Gammaproteobacteria; Enterobacter iales; Enterobacteriaceae; Enterobacter
+xxxxxx	Bacteria; Proteo bacteria; Gammaproteobacteria; Pseudomonadales; Pseudomonadaceae; Pseudomonas
+AB004748	Bacteria; Proteo bacteria; Gammaproteobacteria; Enterobacter iales; Enterobacteriaceae; Enterobacter
+AB000278	Bacteria; Proteo bacteria; Gammaproteobacteria; Vibr_iona_les; Vibrionaceae; Photobacterium
+AB000390	Bacteria; Proteo bacteria; Gammaproteobacteria; Vibr iona les; Vibrionaceae; Vibrio
+"""
+
+mothur_input_collision = """\
+>underscores_seq
+CAGGCCTAACACATGCAAGTCGAACGGTAANAGATTGATAGCTTGCTATCAATGCTGACGANCGGCGGACGGGTGAGTAATGCCTGGGAATATACCCTGATGTGGGGGATAACTATTGGAAACGATAGCTAATACCGCATAATCTCTTCGGAGCAAAGAGGGGGACCTTCGGGCCTCTCGCGTCAGGATTAGCCCAGGTGGGATTAGCTAGTTGGTGGGGTAATGGCTCACCAAGGCGACGATCCCTAGCTGGTCTGAGAGGATGATCAGCCACACTGGAACTGAGACACGGTCCAGACTCCTACGGGAGGCAGCAGTGGGGAATATTGCACAATGGGGGAAACCCTGATGCAGCCATGCCGCGTGTA
+>spaces_seq
+TGGCTCAGATTGAACGCTGGCGGCAGGCCTAACACATGCAAGTCGAGCGGAAACGANTNNTNTGAACCTTCGGGGNACGATNACGGCGTCGAGCGGCGGACGGGTGAGTAATGCCTGGGAAATTGCCCTGATGTGGGGGATAACTATTGGAAACGATAGCTAATACCGCATAATGTCTACGGACCAAAGAGGGGGACCTTCGGGCCTCTCGCTTCAGGATATGCCCAGGTGGGATTAGCTAGTTGGTGAGGTAATGGCTCACCAAGGCGACGATCCCTAGCTGGTCTGAGAGGATGATCAGCCACACTGGAACTGAG
 """
 
 rdp_reference_seqs = \